requirements.txt
autopep8
rshell
micropython-phew